+++ packages/flutter_meedu/lib/src/state/listeners/provider_listener.dart
@@ -91,7 +91,7 @@ class _ProviderListenerState<T extends BaseNotifier>
     // check if the onAfterFirstLayout callback needs to be called
     if (widget.onAfterFirstLayout != null) {
       // wait after first frame
-      WidgetsBinding.instance.addPostFrameCallback((_) {
+      WidgetsBinding.instance.endOfFrame.then((_) {
         if (mounted) {
           widget.onAfterFirstLayout!(context, _notifier);
         }