+++ packages/meedu/lib/src/notifiers/base_notifier.dart
@@ -19,7 +19,7 @@ abstract class BaseNotifier<T> {
   }
 
   @mustCallSuper
-  void dispose() {
+  FutureOr<void> dispose() {
     _disposed = true;
   }