+++ packages/meedu/lib/src/notifiers/simple_notifier.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:meta/meta.dart' show mustCallSuper;
 
 import 'base_notifier.dart';
@@ -19,8 +21,8 @@ abstract class SimpleNotifier extends BaseNotifier
   /// use to listen when the controller was deleted from memory
   @override
   @mustCallSuper
-  void dispose() async {
+  FutureOr<void> dispose() {
     super.dispose();
-    await clearListeners();
+    clearListeners();
   }
 }